Factor out ConversionRequest::Options to options.h
Factor out the ConversionRequest::Options structure and its dependent
enums into a new header file options.h, while maintaining full
backward compatibility.

PiperOrigin-RevId: 923211193
